--- a/.cargo/config.toml
+++ b/.cargo/config.toml
@@ -15,6 +15,8 @@ rustflags = ["-C", "link-arg=-fuse-ld=lld"]
 
 [target.'cfg(all())']
 rustflags = [
+  "--cfg",
+  "tokio_unstable",
   "-D",
   "clippy::all",
   "-D",