--- a/runtime/tokio_util.rs
+++ b/runtime/tokio_util.rs
@@ -1,5 +1,46 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
+use core::future::Future;
+use core::marker::PhantomData;
+use tokio::runtime::Runtime;
+use tokio::task::JoinHandle;
+
+pub struct LocalRuntime {
+  inner: Runtime,
+  _not_send: PhantomData<*mut ()>,
+}
+
+impl LocalRuntime {
+  pub fn new() -> Self {
+    Self {
+      inner: create_basic_runtime(),
+      _not_send: PhantomData,
+    }
+  }
+
+  pub fn spawn<F>(&self, fut: F) -> JoinHandle<F::Output>
+  where
+    F: 'static + Future + Send,
+    // We could eliminate this one if necessary, but it's tedious to do so.
+    F::Output: 'static + Send,
+  {
+    // SAFETY: The `LocalRuntime` type is neither Send nor Sync, so it's
+    // not possible to move it across threads. The futures inside it are
+    // only ever polled inside `block_on` calls on the runtime, so since
+    // those calls must happen on the same thread, the futures are not
+    // polled no the wrong thread.
+    //
+    // Note that `Handle::block_on` does not execute tasks on a current-
+    // thread runtime, so you can't use it to poll tasks from another
+    // thread.
+    unsafe { self.inner.spawn(fut) }
+  }
+
+  pub fn block_on<F: Future>(&self, fut: F) -> F::Output {
+    self.inner.block_on(fut)
+  }
+}
+
 pub fn create_basic_runtime() -> tokio::runtime::Runtime {
   tokio::runtime::Builder::new_current_thread()
     .enable_io()
@@ -14,11 +55,63 @@ pub fn create_basic_runtime() -> tokio::runtime::Runtime {
     .unwrap()
 }
 
+pub fn run_local2<F, R>(future: F) -> R
+where
+  F: std::future::Future<Output = R> + 'static,
+  F::Output: Send + 'static,
+{
+  let local = LocalRuntime::new();
+  let join_handle =
+    local.spawn(unsafe { make_me_send::MakeMeSend::new(future) });
+  local.block_on(async move { join_handle.await }).unwrap()
+}
+
+pub fn run_local3<F>(future: F)
+where
+  F: std::future::Future + 'static,
+  F::Output: Send + 'static,
+{
+  let local = LocalRuntime::new();
+  local.spawn(unsafe { make_me_send::MakeMeSend::new(future) });
+}
+
 pub fn run_local<F, R>(future: F) -> R
 where
   F: std::future::Future<Output = R>,
 {
-  let rt = create_basic_runtime();
-  let local = tokio::task::LocalSet::new();
-  local.block_on(&rt, future)
+  let local = LocalRuntime::new();
+  local.block_on(future)
+}
+
+pub mod make_me_send {
+  use core::future::Future;
+  use core::pin::Pin;
+  use core::task::Context;
+  use core::task::Poll;
+
+  pub struct MakeMeSend<F> {
+    future: F,
+  }
+
+  impl<F> MakeMeSend<F> {
+    /// SAFETY: You must ensure that the future is not used
+    /// on the wrong thread.
+    pub unsafe fn new(future: F) -> Self {
+      Self { future }
+    }
+  }
+
+  unsafe impl<F> Send for MakeMeSend<F> {}
+
+  impl<F: Future> Future for MakeMeSend<F> {
+    type Output = F::Output;
+
+    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<F::Output> {
+      unsafe {
+        let me = Pin::into_inner_unchecked(self);
+        let future = Pin::new_unchecked(&mut me.future);
+        future.poll(cx)
+      }
+    }
+  }
 }