--- a/runtime/tokio_util.rs
+++ b/runtime/tokio_util.rs
@@ -51,6 +51,8 @@ pub fn create_basic_runtime() -> tokio::runtime::Runtime {
     // The default value is 512, which is an unhelpfully large thread pool. We
     // don't ever want to have more than a couple dozen threads.
     .max_blocking_threads(32)
+    .global_queue_interval(4096)
+    .event_interval(1024)
     .build()
     .unwrap()
 }
@@ -61,6 +63,21 @@ where
   F::Output: Send + 'static,
 {
   let local = LocalRuntime::new();
+
+  local.spawn(async move {
+    let handle = tokio::runtime::Handle::current();
+    let runtime_monitor = tokio_metrics::RuntimeMonitor::new(&handle);
+
+    // print runtime metrics every 500ms
+    let frequency = std::time::Duration::from_millis(500);
+    tokio::spawn(async move {
+      for metrics in runtime_monitor.intervals() {
+        println!("Metrics = {:?}", metrics);
+        tokio::time::sleep(frequency).await;
+      }
+    });
+  });
+
   let join_handle =
     local.spawn(unsafe { make_me_send::MakeMeSend::new(future) });
   local.block_on(async move { join_handle.await }).unwrap()