--- a/serde_v8/serializable.rs
+++ b/serde_v8/serializable.rs
@@ -35,6 +35,8 @@ pub enum SerializablePkg {
   Serializable(Box<dyn Serializable>),
 }
 
+unsafe impl Send for SerializablePkg {}
+
 impl SerializablePkg {
   pub fn to_v8<'a>(
     &mut self,