Opening for visibility, changes from this PR should be landed in bite-sized chunks separately.

Async deferred ops benchmark:

`v1.33.2`

```
deno run --unstable -A ./cli/bench/async_ops_deferred.js
time 1479 ms rate 676132
time 1451 ms rate 689179
time 1446 ms rate 691562
time 1443 ms rate 693000
time 1443 ms rate 693000
time 1443 ms rate 693000
```
Flamegraph: https://share.firefox.dev/42elIEy

`this branch`

```
time 886 ms rate 1128668
time 861 ms rate 1161440
time 853 ms rate 1172332
time 894 ms rate 1118568
time 852 ms rate 1173708
time 853 ms rate 1172332
time 852 ms rate 1173708
```
Flamegraph: https://share.firefox.dev/3LZiEVZ