@mmastrac 2023-05-09
Partially superseded by https://github.com/denoland/deno/pull/19056