@bartlomieju 2023-05-14
Closing in favor of https://github.com/denoland/deno/pull/19056